Xac nhan user khi dang nhap

+++ database/migrations/2021_04_21_015820_create_coupons_table.php
@@ -17,7 +17,7 @@ class CreateCouponsTable extends Migration
             $table->id();
             $table->string('code')->unique();
             $table->enum('type',['fixed','percent'])->default('fixed');
-            $table->decimal('value',20,2);
+            $table->integer('value');
             $table->enum('status',['active','inactive'])->default('inactive');
             $table->timestamps();
         });
